Refactor module names
